+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - '-p' flag to prompt on all task changing commands
 - Added "addx" command to add completed tasks
 - Added "help" with search terms
+- Added "repeat" command to repeat a task
+- Added description of todo.txt in "help todo.txt"
 
 ## [0.4.0] - 2018-04-18
 ### Changed